// src/app/jsPDF.d.ts
export interface AutoTableStyles {
    cellPadding?: number;
    fontSize?: number;
    font?: string;
    lineColor?: number | number[];
    lineWidth?: number;
    lineHeight?: number;
    fontStyle?: string;
    fillColor?: number | number[];
    textColor?: number | number[];
    halign?: 'left' | 'center' | 'right';
    valign?: 'top' | 'middle' | 'bottom';
    fillStyle?: 'S' | 'F' | 'DF';
    rowHeight?: number;
    columnWidth?: 'auto' | 'wrap' | number;
    overflow?: 'linebreak';
}

export interface AutoTableRow {
    raw: any;
    index: number;
    cells: any;
    spansMultiplePages: boolean;
    pageCount: number;
    height: number;
    y: number;
    maxLineCount: number;
}

export interface AutoTableCell {
    raw: any;
    styles: AutoTableStyles;
    text: string;
    contentWidth: number;
    textPos: { x: number, y: number };
    height: number;
    width: number;
    x: number;
    y: number;
}

export interface AutoTableData {
    pageCount: number;
    settings: any;
    table: any;
    doc: jsPDF;
    cursor: { x: number, y: number };
}